Convert svelte modules to plain js.
